fix: reinstall prettier
