+++ class-component/src/apiUtils.ts
@@ -1,0 +1,11 @@
+export const fetchFilmTitle = async (url: string): Promise<string> => {
+    const response = await fetch(url);
+    const data = await response.json();
+    return data.title;
+};
+
+export const fetchHomeworldName = async (url: string): Promise<string> => {
+    const response = await fetch(url);
+    const data = await response.json();
+    return data.name;
+};
